chore: Remove old Quarto artifacts, regenerate with hook
Deleted stale generated files. Pre-commit hook will regenerate fresh artifacts with current code.
